requirements.txt
-i https://pypi.org/simple
aioconsole==0.7.0
aiohttp==3.9.1
aiosignal==1.3.1; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7
colorama==0.4.6; platform_system == 'Windows'
coloredlogs==15.0.1
dictor==0.1.12
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord-py @ git+https://github.com/Rapptz/discord.py@50190e088e38257016e69b9b0004f17b814b4478
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
frozenlist==1.4.0; python_version >= '3.8'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.6; python_version >= '3.5'
motor==3.3.2
multidict==6.0.4; python_version >= '3.7'
psutil==5.9.6
pymongo==4.6.1
pyreadline3==3.4.1; python_version >= '3.8' and sys_platform == 'win32'
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
sentry-sdk==1.39.0
urllib3==2.1.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
